streamlit/requirements.txt
# ──────────────────────────────────────────────
#  Streamlit frontend
# ──────────────────────────────────────────────
streamlit==1.55.0

# ──────────────────────────────────────────────
#  HTTP — calls the FastAPI backend
# ──────────────────────────────────────────────
requests==2.32.5

# ──────────────────────────────────────────────
#  Computer vision & image processing
# ──────────────────────────────────────────────
opencv-python-headless==4.10.0.84   # headless = no GUI deps
Pillow==10.2.0
numpy==1.26.4